Más tests, no todos correctos, algunos presentan fallos

+++ Back/src/main/java/es/cic/controlaereo/Back/servicio/NaveServicio.java
@@ -41,6 +41,8 @@ public class NaveServicio {
     }
 
     public void deleteNaveById(Long id) {
-        naveRepositorio.deleteById(id);
+        if (naveRepositorio.existsById(id)) {
+            naveRepositorio.deleteById(id);
+        }
     }
 }
